+++ source/platforms/bitbucket_cloud/_tests_/_bitbucket_cloud_git.test.ts
@@ -51,6 +51,7 @@ describe("the dangerfile gitDSL - BitBucket Cloud", async () => {
       username: "username",
       password: "password",
       uuid: "{1234-1234-1234-1234}",
+      type: "PASSWORD",
     })
     bbc = new BitBucketCloud(api)